+++ spec/arachni/rpc/server/instance_spec.rb
@@ -54,7 +54,7 @@ describe Arachni::RPC::Server::Instance do
             end
         end
 
-        describe '#configure_and_scan' do
+        describe '#scan' do
             it 'should configure and start a scan' do
                 instance = @get_instance.call
 
@@ -63,13 +63,16 @@ describe Arachni::RPC::Server::Instance do
                 instance.service.busy?.should  == instance.framework.busy?
                 instance.service.status.should == instance.framework.status
 
-                instance.service.configure_and_scan(
+                instance.service.scan(
                     url:         server_url_for( :framework_simple ),
                     audit_links: true,
                     audit_forms: true,
                     modules:     :test,
                     slaves:      [ { url: slave.url, token: @token } ]
-                )
+                ).should be_true
+
+                # if a scan in already running it should just bail out early
+                instance.service.scan.should be_false
 
                 sleep 1 while instance.service.busy?
 
@@ -83,6 +86,51 @@ describe Arachni::RPC::Server::Instance do
 
                 i_report.should == f_report
                 i_report['issues'].should be_any
+            end
+        end
+
+        describe '#progress' do
+            it 'should return progress information' do
+                instance = @get_instance.call
+
+                slave = @get_instance.call
+
+                p = instance.service.progress
+
+                p['busy'].should   == instance.framework.busy?
+                p['status'].should == instance.framework.status
+                p['stats'].should  == instance.framework.progress['stats']
+
+                p['instances'].should be_nil
+                p['messages'].should be_nil
+                p['issues'].should be_nil
+
+                instance.service.progress( true )['issues'].should be_empty
+
+                instance.service.scan(
+                    url:         server_url_for( :framework_simple ),
+                    audit_links: true,
+                    audit_forms: true,
+                    modules:     :test,
+                    slaves:      [ { url: slave.url, token: @token } ]
+                ).should be_true
+
+                sleep 1 while instance.service.busy?
+
+                p = instance.service.progress
+
+                p['busy'].should   == instance.framework.busy?
+                p['status'].should == instance.framework.status
+                p['stats'].should  == instance.framework.progress['stats']
+
+                p['instances'].size.should == 2
+                p['instances'].should == instance.framework.progress_data['instances']
+
+                p['messages'].should be_nil
+
+                issues = instance.service.progress( true )['issues']
+                issues.should be_any
+                issues.should == instance.framework.progress_data( as_hash: true )['issues']
             end
         end